+++ src/client/app/app-routing.module.ts
@@ -12,4 +12,3 @@ import { RouterModule } from '@angular/router';
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
-