+++ src/client/app/header/header-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { HeaderComponent } from './header.component';
+
+@NgModule({
+  imports: [
+    RouterModule.forChild([
+      { path: '**', component: HeaderComponent }
+    ])
+  ],
+  exports: [RouterModule]
+})
+export class HeaderRoutingModule { }