+++ src/client/app/header/header.module.ts
@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { HeaderRoutingModule } from './header-routing.module';
 import { HeaderComponent } from './header.component';
 
 @NgModule({
-    imports: [CommonModule],
+    imports: [HeaderRoutingModule, CommonModule],
     declarations: [HeaderComponent],
     exports: [HeaderComponent]
 })